Move lambda example to the main integration tests
