+++ integration-tests/main/src/main/java/io/quarkus/example/lambda/HelloGreeter.java
@@ -1,4 +1,4 @@
-package io.quarkus.lambda.sample;
+package io.quarkus.example.lambda;
 
 import javax.enterprise.context.ApplicationScoped;
