+++ integration-tests/main/src/test/java/io/quarkus/example/test/LambdaITCase.java
@@ -1,4 +1,4 @@
-package io.quarkus.lambda.sample;
+package io.quarkus.example.test;
 
 import io.quarkus.test.junit.SubstrateTest;
