+++ integration-tests/pom.xml
@@ -51,6 +51,5 @@
         <module>infinispan-cache-jpa</module>
         <module>infinispan-cache-jpa-stress</module>
         <module>panache</module>
-        <module>lambda</module>
     </modules>
 </project>
